[sw/device,tests] Add sim_verilator environment for different versions of bnmulv{l}
This commit modifies the bazel rules so that fusesoc knows which version
of the RTL to simulate. In particular:
- fusesoc.bzl: fusesoc_build rule now takes "bnmulv_version_id" as its
attribute. This attribute will be added directly in fusesoc args as
"--flag bnmulv_ver*" in fusesoc_build_impl.
- hw/BUILD: we addd fusesoc_build targets "verilator_real_ver*" (alias
"verilator_ver*").
- hw/top_earlgrey/BUILD: "verilator_ver*" is used to build verilator
environment "sim_verilator_ver*" for top_earlgrey.
- defs.bzl: //hw/top_earlgrey:sim_verilator_ver* is added in EARLGREY_TEST_ENVS,
which is used in sw/device/tests.
Each version has its own environment because //hw:verilator_ver* is built first
when opentitan_test is called and it doesn't recognized any options passed
in the command line. opentitan_test just loads the software on the built env
and runs the test.

In order for the chip to recognize the fusesoc flags bnmulv_ver*, we need
to add "bnmulv_ver* ? (-DBNMULV_VER*)" in hw/top_earlgrey/dv/verilator/chip_sim.core.

On the software side, we modified/added:
- otbn_as: it now puts all "-D" options (either passed through the commands or
in the BUILD targets) in one list. The first -D option is always the one declared
in BUILD targets. The rest is either "-DBNMULV_VER=*" or "-DKYBER_K/-DDILITHIUM_MODE"
which are both passed in the command line when running the FPGA tests. Note that
this change is very specific to how BUILD targets for MLKEM and MLDSA are defined.
- sw/device/tests: Add chip-level tests for MLKEM and MLDSA. To run the tests:
./bazelisk.sh test --test_timeout=10000 --test_output=streamed --copt="-DBNMULV_VER=*" --copt="-DKYBER_K=*/DILITHIUM_MODE=*" //sw/device/tests:otbn_mlkem_test_sim_verilator_ver*
WARNING: sim_verilator_ver* tag must match BNMULV_VER*
NOTE: ML-KEM tests passed! ML-DSA tests are not tested with Verilator
because they take a lot of time.