[hw/bitstream] Add bazel targets for building bitstream CW310 depending on BNMULV_VER
This commit hasn't been tested yet. It is put here to be tested on
the server with proper Xilinx Vivado license. But ideally, it would
go as follows:
- Build the bitstream for CW310 manually:
./bazelisk.sh build //hw/bitstream/vivado:fpga_cw310_test_rom{_ver1,_ver2,_ver3}
- Load the bitstream onto the FPGA:
./bazelisk.sh run //sw/host/opentitantool -- fpga load-bitstream /path/to/bitstream
- Run the test:
./bazelisk.sh test --define bistream=skip --test_output=streamed \
--copt="-DBNMULV_VER=*" --copt="-DKYBER_K=*/-DDILITHIUM_MODE=*" \
//sw/device/tests:otbn_{mlkem,mldsa}_test_fpga_cw310_test_rom{_ver1,_ver2,_ver3}
The --define bitstream=skip is because we already loaded the bitstream onto the FPGA.
If we want to directly build the bitstream with this command, replace
"--define bitstream=skip" by "--define bistream=vivado".