[otbn,mlkem] Remove cond sub in VER2
NTT is used in the following parts:
- In keypair, NTT(skpv) --> pack_sk
- In keypair, NTT(e) --> pk + e --> pack_pk
- In encap, NTT(sp) --> basemul: v = sp * pkpv
- In decap, NTT(b) --> basemul: v = v * b
If we remove conditional subtraction in Montgomery multiplication in NTT,
the coefficient can grow to 2q. Then when it is subtracted from another
coeff in [0,q), the result can be negative, represented as (mod 2**16).
The modular multiplications with this result from then on will be wrong,
because of this extra mod 2**16 term. So in order to deal with this, we
switch the modulus when appropriate.

The input of NTT is [0,q). After the first layer, b*root isi n [0,2q).
The subtraction a - b*root is in (-2q,q) and addition a + b*root is in
[0,3q). With addvm and MOD = q, the results are (-q,q) and [0,2q) respectively.
This -q is the problem. Now, if we switch MOD = 2q, the results would be
in [0,2q) and [0,2q). Then from the second layer, when a and b*root are both
in [0,2q), their sum and subtraction would be in [0,4q) and (-2q,2q).
With addvm and MOD = 2q, they would be [0,2q) and [0,2q) respectively.
The result of NTT would be in [0,2q).

Thus, the idea is to switch MOD to 2q before NTT and switch it back to q
once done.

For encap/decap, after NTT, basemul computes a0*b0 + a1*b1*root and a0*b1 + a1*b0.
It doesn't matter if the inputs are in [0,2q) because the multiplications
are with conditional subtraction in the end, ensuring the results to be in [0,q).
For keypair, after NTT(skpv) is pack_sk, which requires inputs to be in [0,q). So
we need addvm for skpv before packing. Similarly to NTT(e). Hence, while
performance of encap/decap is improved, that of keypair is worse, but only
maximum of 221 cycles for mlkem1024, since we need to the coeffs again
and addvm them with 0 then save them back to the memory. In theory, the perf
of keypair is also improved if we put addvms inside the NTT, but these are
not necessary for basemul. Thus, to reduce perf of encap/decap and keeps
the same code size, we sacrifice perf of keypair, but very little. The most
important result from this is we have shown that cond_sub VER3 is not needed.

INTT is used the following parts:
- In encap, INTT(v) --> pack_ct
- In encap, INTT(b) --> pack_ct
- In decap, INTT(b) --> v = v - b

After basemul, inputs to INTT are in [0,q). After the first layer,
a + b and a - b will be in [0,2q) and (-q,q) respectively. Then with
addvm and subvm, they'll be in [0,q) and [0,q), which is still fine.
If the final cond sub in Montgomery multiplication is removed, then
(a - b)*root will be in [0,2q). Then in the second layer, a + b
can either be [0,2q) if both are in [0,q) or [0,4q) if both are in
[0,2q). (a - b) will be either (-q,q) if both are in [0,q) or (-2q,2q)
if both are in [0,2q). With addvm, a + b will be in [0,q) or [0,3q);
with subvm, a - b will be in [0,q) or [-q,2q). This "-q" is the problem.

So similarly to NTT, we switch MOD to 2q before entering INTT. Then,
after the first layer, a + b (addvm) will be in [0,2q) and (a - b)
(subvm) will be in [0,2q). From the second layer onwards, both a + b
(addvm) and a - b (subvm) will still be in [0,2q). Hence, the output
of INTT would be in [0,2q).

Now in the last layer of INTT, half of the coeffs are multiplied with
root of unity and "1/128 mod q" at the same time. Then the other half
is multiplied with "1/128 mod q" separately. The idea is to switch MOD
back to q so Montgomery multiplications with "1/128 mod q" are kept
unchanged (final cond_sub is not removed to put coeffs in [0,q)). For
the first half, since it is inefficient to switch MOD back and forth
between q and 2q in the last layer, we add 8 addvms with 0 and MOD = q
to reduce the coeffs in [0,q). Then we switch MOD back to 2q before
exiting INTT.

In basemul, originally, we compute the following products:
- a0*b0 --> [0,q)
- a1*b1 --> [0,q)
- (a1*b1)*root --> [0,q)
- a0*b1 --> [0,q)
- a1*b0 --> [0,q)
Then we compute the following sums:
- a0*b0 + a1*b1*root --> [0,q)
- a0*b1 + a1*b0 --> [0,q)

Now if we remove final conditional subtraction after Montgomery
multiplication, we have:
- a0*b0, a1*b1*root, a1*b0, a0*b1 in [0,2q)
And thus the respective sums would be in [0,3q). After accumulation
for K times, the output of vector-vector product can grow to [0,9q),
which is too much for INTT.

So the idea is to switch MOD to 2q as in INTT/NTT and remove final
conditional subtraction. Then:
- a0*b0 + a1*b1*root --> [0,2q)
- a1*b0 + a0*b1 --> [0,2q)
- Accumulation: [0,2q) + [0,2q) --> [0,2q)
In the end, the outputs of vector-vector product would be in [0,2q).
For encap/decap, this is input to INTT with MOD = 2q, which is fine.
For keypair, this is input to poly_tomont with MOD = q, which means
the results are reduced mod q, correct.